requirements.txt
numpy>=1.20,<2.0        # array ops for env & policies
requests>=2.31          # HTTPS calls to Deepseek & Qwen (or OpenRouter)
python-dotenv>=1.0      # loads .env into `config.py`
matplotlib>=3.8         # plotting Figure 2 style cost-savings curve
